portfolio-optimizer/requirements.txt
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
yfinance